ashamed to commit this garbage

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: hfegrach <hfegrach@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 17:30:51 by hfegrach          #+#    #+#             */
-/*   Updated: 2024/11/10 15:12:45 by hfegrach         ###   ########.fr       */
+/*   Updated: 2025/01/25 02:46:54 by hfegrach         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,11 +36,11 @@ int	ft_atoi(const char *a)
 	}
 	return ((int)(sign * res));
 }
-long	ft_atol(const char *a)
+long long	ft_atol(const char *a)
 {
 	int		i;
 	int     sign;
-	long	res;
+	long long	res;
 
 	i = 0;
 	sign = 1;
